requirements.txt
autopep8==1.5.4
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
chromedriver-binary==84.0.4147.30.0
cycler==0.10.0
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
kiwisolver==1.2.0
matplotlib==3.3.0
mccabe==0.6.1
numpy==1.19.1
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
